requirements.txt
Pillow==9.2.0
django-taggit==3.0.0
django-taggit-templatetags2==1.6.1
django-widget-tweaks==1.4.12